Move WebViewDragOverlay into its own file.
