+++ v3/src/components/web-view/web-view-drop-overlay.scss
@@ -1,0 +1,5 @@
+.codap-web-view-drop-overlay {
+  height: 100%;
+  position: absolute;
+  width: 100%;
+}
