+++ v3/src/components/web-view/web-view.scss
@@ -40,10 +40,4 @@
       width: 100%;
     }
   }
-
-  .codap-web-view-drop-overlay {
-    height: 100%;
-    position: absolute;
-    width: 100%;
-  }
 }
