+++ v3/src/components/web-view/web-view.tsx
@@ -1,12 +1,9 @@
-import { Active, useDroppable } from "@dnd-kit/core"
 import { observer } from "mobx-react-lite"
-import React, { MouseEventHandler, useRef } from "react"
-import { getDragAttributeInfo, useDropHandler } from "../../hooks/use-drag-drop"
-import { useTileModelContext } from "../../hooks/use-tile-model-context"
-import { dragNotification, dragWithPositionNotification } from "../../lib/dnd-kit/dnd-notifications"
+import React, { useRef } from "react"
 import { t } from "../../utilities/translation/translate"
 import { ITileBaseProps } from "../tiles/tile-base-props"
 import { useDataInteractiveController } from "./use-data-interactive-controller"
+import { WebViewDropOverlay } from "./web-view-drop-overlay"
 import { isWebViewModel } from "./web-view-model"
 
 import "./web-view.scss"
@@ -34,64 +31,3 @@ export const WebViewComponent = observer(function WebViewComponent({ tile }: ITi
     </div>
   )
 })
-
-// The WebViewDropOverlay broadcasts notifications to plugins as the user drags and drops attributes.
-function WebViewDropOverlay() {
-  const overlayRef = useRef<HTMLElement|null>()
-  const { tile, tileId } = useTileModelContext()
-  const dropId = `web-view-drop-overlay-${tileId}`
-  const { active, setNodeRef } = useDroppable({ id: dropId })
-  const info = active && getDragAttributeInfo(active)
-  const dataSet = info?.dataSet
-  const attributeId = info?.attributeId
-  // Mouse x and y are tracked so we know where the mouse is when a dragged attribute is dropped.
-  const mouseX = useRef<number|undefined>()
-  const mouseY = useRef<number|undefined>()
-
-  useDropHandler(dropId, (_active: Active) => {
-    const { dataSet: dropDataSet, attributeId: dropAttributeId } = getDragAttributeInfo(_active) || {}
-    if (dropDataSet && dropAttributeId && mouseX.current != null && mouseY.current != null) {
-      tile?.applyModelChange(() => {}, {
-        notify: dragWithPositionNotification("drop", dropDataSet, dropAttributeId, mouseX.current, mouseY.current),
-        notifyTileId: tileId
-      })
-    }
-  })
-
-  if (!dataSet || !attributeId) return null
-
-  const handleMouseMove: MouseEventHandler<HTMLDivElement> = event => {
-    const { top, left } = overlayRef.current?.getBoundingClientRect() ?? { top: 0, left: 0 }
-    const x = event.clientX - left
-    const y = event.clientY - top
-
-    if (mouseX.current !== x || mouseY.current !== y) {
-      tile?.applyModelChange(() => {}, {
-        notify: dragWithPositionNotification("drag", dataSet, attributeId, x, y),
-        notifyTileId: tileId
-      })
-      mouseX.current = x
-      mouseY.current = y
-    }
-  }
-
-  const handleMouseEnterLeave = (operation: string) => {
-    tile?.applyModelChange(() => {}, {
-      notify: dragNotification(operation, dataSet, attributeId),
-      notifyTileId: tileId
-    })
-  }
-
-  const setRef = (ref: HTMLDivElement) => {
-    overlayRef.current = ref
-    setNodeRef(ref)
-  }
-
-  return <div
-    className="codap-web-view-drop-overlay"
-    onMouseEnter={() => handleMouseEnterLeave("dragenter")}
-    onMouseLeave={() => handleMouseEnterLeave("dragleave")}
-    onMouseMove={handleMouseMove}
-    ref={setRef}
-  />
-}
